+++ packages/react-components/src/Automations/helper/const.ts
@@ -1,3 +1,4 @@
+import { JSONSchema7 } from "json-schema";
 
 export const fields = ['name', 'group', 'taskId', 'hostGroup', 'tag'];
 export const triggerFields = ['triggerCondition'];
@@ -31,4 +32,43 @@ export const initialTriggerError = {
   triggerConditionError: false,
   triggerIdError: false,
   typeError: false,
+};
+
+export const uiSchema = {
+  startTimeUtc: {
+    "ui:options": {
+      label: false
+    }
+  },
+  description: {
+    "ui:widget": "text",
+    "ui:options": {
+      label: false
+    }
+  },
+  interval: {
+    "ui:widget": "text",
+    "ui:options": {
+      label: false
+    }
+  },
+};
+
+export const schema : JSONSchema7 = {
+  type: "object",
+  properties: {
+    startTimeUtc: {
+      type: "string",
+      format: "date-time",
+      title: "Start Date Time"
+    },
+    description: {
+      type: "string",
+      title: "Description"
+    },
+    interval: {
+      type: "string",
+      title: "Interval"
+    }
+  }
 };
